refactor(leaderboard): isolate secure CSV utilities and add regression tests
