+++ package.json
@@ -9,7 +9,8 @@
     "lint": "eslint .",
     "preview": "vite preview",
     "cf:dev": "npx wrangler dev --config cloudflare/wrangler.toml",
-    "cf:deploy": "npx wrangler deploy --config cloudflare/wrangler.toml"
+    "cf:deploy": "npx wrangler deploy --config cloudflare/wrangler.toml",
+    "test": "vitest run"
   },
   "dependencies": {
     "react": "^19.2.4",
@@ -29,6 +30,7 @@
     "globals": "^17.4.0",
     "postcss": "^8.5.8",
     "tailwindcss": "^3.4.19",
-    "vite": "^8.0.1"
+    "vite": "^8.0.1",
+    "vitest": "^4.1.7"
   }
 }
